## Changed: - moved fetching data logic to shared component
